Tiny changes and added statistics for memory allocations.
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@102 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/angelscript/source/as_string.cpp
@@ -120,7 +120,7 @@ void asCString::Allocate(size_t len, bool keepData)
 
 		if( keepData )
 		{
-			int l = len < length ? len : length;
+			int l = (int)len < length ? (int)len : length;
 			memcpy(buf, AddressOf(), l);
 		}
 
@@ -144,7 +144,7 @@ void asCString::Allocate(size_t len, bool keepData)
 		}
 	}
 
-	length = len;
+	length = (int)len;
 
 	// Make sure the buffer is null terminated
 	AddressOf()[length] = 0;
